tests refactored
