+++ src/test/scala/closures/mocks/MockCreateNewPeriodClosure.scala
@@ -1,0 +1,12 @@
+package closures.mocks
+
+import closures.CreateNewPeriodClosure
+import dao.hbase.MockCreateNewPeriodHBaseDao
+
+import scala.util.Random
+
+object MockCreateNewPeriodClosure extends CreateNewPeriodClosure{
+  override val hbaseDao = MockCreateNewPeriodHBaseDao
+  override def generateUniqueKey = Random.alphanumeric.take(12).mkString + "TESTS" //to ensure letters and numbers present
+
+}
