+++ src/test/scala/test/data/existing/ExistingData.scala
@@ -1,3 +1,3 @@
-package dao.parquet
+package test.data.existing
 
 trait ExistingData extends ExistingEnts with ExistingLinks with ExistingLocalUnits
